fixing types in types.hpp header

+++ Core/src/cpp/main/gas/types.hpp
@@ -9,10 +9,10 @@ namespace gas{
     using char_p = char_t*;
     using char_cp = const char_t*;
     
-    using u8 = uint8_t;    
-    using u8 = uint8_t;    
-    using u8 = uint8_t;    
-    using u8 = uint8_t;
+    using u8  = uint8_t;    
+    using u16 = uint16_t;    
+    using u32 = uint32_t;    
+    using u64 = uint64_t;
 }
 
 #endif
